examples/example-project/pyproject.toml
[tool.poetry]
name = "example-project"
version = "0.1.0"
description = ""
authors = ["Gustavo Correa <self.gustavocorrea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.18.0"
aiomysql = "^0.0.21"
sqlalchemy = "^1.4.21"
fastapi = "^0.66.0"
pydantic = { extras = ["email"], version= "^1.8.2" }
circus = "^0.17.1"
gunicorn = "^20.1.0"
aiohttp = "^3.7.4.post0"
uvicorn = "^0.14.0"
uvloop = { version= "^0.15.3", optional = true }
httptools = { version= "^0.2.0", optional = true }

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pytest-cov = "^2.12.1"
pytest = "^6.2.4"
pylint = "^2.9.3"
sqlalchemy2-stubs = "^0.0.2a4"
pytest-asyncio = "^0.15.1"

[tool.poetry.scripts]
start = "example_project.main:main"

[tool.poetry.extras]
deploy = ["httptools", "uvloop"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

